Crate sc_network_types

Modules

ed25519

Ed25519 keys.

multiaddr

multihash

Multihash implemenattion used by substrate. Currently it’s a wrapper over multihash used by litep2p, but it can be switched to other implementation if needed.

Macros

build_multiaddr

Easy way for a user to create a Multiaddr.

Structs

PeerId

Identifier of a peer of the network.