Struct sc_network_types::ed25519::SecretKey

pub struct SecretKey(/* private fields */);

An Ed25519 secret key.

Implementations

impl SecretKey

pub fn generate() -> SecretKey

Generate a new Ed25519 secret key.

pub fn try_from_bytes( sk_bytes: impl AsMut<[u8]> ) -> Result<SecretKey, DecodingError>

Try to parse an Ed25519 secret key from a byte slice containing the actual key, zeroing the input on success. If the bytes do not constitute a valid Ed25519 secret key, an error is returned.

pub fn to_bytes(&self) -> [u8; 32]