sc_network_types::multiaddr

Struct Multiaddr

pub struct Multiaddr { /* private fields */ }

Multiaddr type used in Substrate. Converted to libp2p’s Multiaddr or litep2p’s Multiaddr when passed to the corresponding network backend.

Implementations

impl Multiaddr

pub fn empty() -> Self

Create a new, empty multiaddress.

pub fn push(&mut self, p: Protocol<'_>)

Adds an address component to the end of this multiaddr.

pub fn pop<'a>(&mut self) -> Option<Protocol<'a>>

Pops the last Protocol of this multiaddr, or None if the multiaddr is empty.

pub fn with(self, p: Protocol<'_>) -> Self

Like Multiaddr::push but consumes self.

pub fn iter(&self) -> Iter<'_>

Returns the components of this multiaddress.

pub fn to_vec(&self) -> Vec<u8>

Return a copy of this Multiaddr’s byte representation.